AMP-61335 more configuration clean up
