+++ packages/cli/src/config/UserConfig.ts
@@ -1,5 +1,10 @@
-import { UserConfigModel } from "../config";
 import { JsonSchemaPropertyModel } from "../json-schema";
+import { PropertyConfigModel } from "./PropertyConfigModel";
+
+export interface UserConfigModel {
+  properties?: Record<string, PropertyConfigModel>;
+  required?: string[];
+}
 
 export class UserConfig {
   constructor(private model: UserConfigModel) {}
